requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
paddleocr>=2.7.0
paddlepaddle>=2.5.2
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.38.0
pyzbar>=0.1.9
opencv-python>=4.9.0.80
Pillow>=10.2.0
pdf2image>=1.17.0
rapidfuzz>=3.6.1
sqlalchemy>=2.0.27
aiofiles>=23.2.1
faker>=22.6.0
numpy>=1.26.4
spacy>=3.7.0
easyocr>=1.7.2
google-cloud-vision>=3.7.0
